:art: Refactor components into own files
